POS api functions
git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@5468 09516a42-71f9-43be-811f-9b821cf0d246

+++ Z_POS_Data.php
@@ -12,8 +12,12 @@ $title = _('Create POS Data Upload File');
 include('includes/header.inc');
 
 if (isset($_GET['Delete'])){
-	unlink($_SESSION['reports_dir'] . '/POS.sql');
-	unlink($_SESSION['reports_dir'] . '/POS.sql.zip');
+	if (file_exists($_SESSION['reports_dir'] . '/POS.sql')){
+		unlink($_SESSION['reports_dir'] . '/POS.sql');
+	}
+	if (file_exists($_SESSION['reports_dir'] . '/POS.sql.zip')){
+		unlink($_SESSION['reports_dir'] . '/POS.sql.zip');
+	}
 	prnMsg(_('Old POS upload files deleted'),'info');
 }
